pyproject.toml
[build-system]
requires = ["setuptools~=68.0.0", "wheel~=0.40.0"]
build-backend = "setuptools.build_meta"

[project]
name              = "dbus_idle"
version           = "2024.3.0"
description       = "System idle time using DBus"
readme            = "README.md"
keywords          = ["lnxlink"]
requires-python   = ">=3.7.0"
authors     = [
    {name="Alexander Frenzel", email="alex@relatedworks.com"},
    {name="bkbilly", email="bkbilly@hotmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
    "dbus-python>=1.3.2"
]


[project.urls]
"Source Code" = "https://github.com/bkbilly/dbus_idle"
"Home Page"   = "https://bkbilly.github.io/dbus_idle"

[tool.setuptools.packages.find]
include = ["dbus_idle*"]

[project.scripts]
dbus-idle = "dbus_idle.__main__:main"
